(PC-29910)[PRO] refactor: Rename NewRequestFormDialog in RequestFormDialog.
